requirements.txt
mysql-connector-python
customtkinter
requests
packaging